# Conflicts:
#	ios/Podfile.lock
#	lib/src/explore/view/screens/product_description_screen.dart
#	lib/src/order/view/widgets/order_view_item_widget.dart
#	pubspec.yaml

+++ lib/src/order/model/create_order.dart
@@ -1,5 +1,5 @@
 class CreateOrder {
-  double? totalAmount;
+  int? totalAmount;
   String? deliveryInstructions;
   String? deliveryType;
   ShippingInfo? shippingInfo;
